requirements.txt
matplotlib
numpy
opencv_python
pandas
Pillow
progressbar33
protobuf
psutil
tensorflow
tensorflow_hub
